Remove `get_model` from model.py to avoid uneccesary import `get_model` in model.py is moved to the test utils as it was not used in gordo-components.
