--- gordo_components/model/__init__.py
@@ -1,1 +1,0 @@
-from .model import get_model
